+++ Solutions/Workplace from Facebook/Data Connectors/WorkplaceFacebook/azuredeploy_Connector_FacebookWorkplaceWebhooks_AzureFunction.json
