+++ Solutions/Workplace from Facebook/Data/Solution_WorkplaceFacebook.json
@@ -1,0 +1,17 @@
+{
+  "Name": "Workplace from Facebook",
+  "Author": "Facebook",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Workbooks/Images/Logos/Azure_Sentinel.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [Workplace](https://www.workplace.com/) data connector provides the capability to ingest common Workplace events into Microsoft Sentinel through Webhooks.\n\n**Underlying Microsoft Technologies used:**\n\nThis solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs:\n\n- [Azure Monitor HTTP Data Collector API](https://docs.microsoft.com/azure/azure-monitor/logs/data-collector-api)\n\n- [Azure Functions](https://azure.microsoft.com/services/functions/#overview)",
+  "Data Connectors": [
+    "DataConnectors/WorkplaceFacebook/WorkplaceFacebook_Webhooks_FunctionApp.json"
+  ],
+  "Parsers": [
+    "Parsers/WorkplaceFacebook/Workplace_Facebook.txt"
+  ],
+  "BasePath": "C:\\Users\\v-eliforbes\\Azure-Sentinel",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+  "TemplateSpec": true,
+  "Is1PConnector": false
+}
