+++ Solutions/Workplace from Facebook/SolutionMetadata.json
@@ -1,0 +1,15 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-workplacefromfacebook",
+	"firstPublishDate": "2022-05-18",
+	"providers": ["Facebook"],
+	"categories": {
+		"domains" : ["Application"]
+	},
+	"support": {
+        "name": "Microsoft Corporation",
+        "email": "support@microsoft.com",
+        "tier": "Microsoft",
+        "link": "https://support.microsoft.com"
+    }
+}
